# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "",
        "version": "0.0.0"
    },
    "paths": {
...
        "/api/users/confirm-email": {
            "post": {
                "operationId": "users_confirm_email_create",
                "description": "Функция подтверждения почты",
                "tags": [
                    "Пользователи"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserConfirmEmailRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/users/get": {
            "get": {
                "operationId": "users_get_retrieve",
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "cookie",
                        "name": "token",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "tags": [
                    "Пользователи"
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 40
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
...
            "UserConfirmEmailRequest": {
                "type": "object",
                "properties": {
                    "confirm-token": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "confirm-token"
                ]
            },
...
        }
    }
}